pyproject.toml
[project]
name = "chesscom"
version = "0.1.0"
description = "Python tools for managing a Chess.com club — member reports, match participation, prospect lists, and match eligibility."
requires-python = ">=3.11"
dependencies = [
    "numpy~=2.4.0",
    "pandas~=2.3.3",
    "requests~=2.32.5",
    "openpyxl~=3.1.5",
    "python-dotenv~=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest~=9.0.2",
    "pytest-cov~=7.0.0",
    "responses~=0.26.0",
    "ruff~=0.15.2",
]

[project.scripts]
chesscom = "chesscom.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501", # line too long — handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["chesscom"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
